Doesn't inherit from MUI button
