+++ web/packages/new/photos/components/mui/index.tsx
@@ -18,31 +18,6 @@ export const FilledIconButton = styled(IconButton)(({ theme }) => ({
 }));
 
 /**
- * An "unstyled" button.
- *
- * There are cases where we semantically (and functionally) want a button, but
- * don't want the browser's default styling. This component is meant to act as a
- * base for such cases.
- *
- * Contrary to its name, it does add a bit of styling, to make these buttons fit
- * in with the rest of our theme.
- */
-export const UnstyledButton = styled("button")`
-    /* Reset some button defaults that are affecting us */
-    background: transparent;
-    border: 0;
-    padding: 0;
-
-    font: inherit;
-    /* We customized the letter spacing in the theme. Need to tell the button to
-       inherit that customization also. */
-    letter-spacing: inherit;
-
-    /* Default cursor on mouse over of a button is not a hand pointer */
-    cursor: pointer;
-`;
-
-/**
  * A flexbox with justify content set to space-between and center alignment.
  *
  * There is also another SpaceBetweenFlex in the old shared package, but that
